Updated the tests.
